requirements.txt
flask==2.2.5
python-dotenv==1.0.1
pandas==2.2.1
ta==0.10.2
pybit-unofficial==3.2.4